pyproject.toml
[tool.poetry]
name = "Flask-Admin-Coloured"
version = "0.1.0"
description = "Adding font and cell-fill colours to Flask-Admin"
authors = ["Sean McCarthy <sean.mccarthy@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
flask-admin = "^1.5.7"
flask-sqlalchemy = "^2.4.4"
flask = "^1.1.2"
pyjwt = "^1.7.1"
python-dotenv = "^0.15.0"
flask-cors = "^3.0.9"
flask-restx = "^0.2.0"
flask-restplus-sqlalchemy = "^0.0.3"
flask-restplus = "^0.13.0"
werkzeug = "0.16.1"
flask-login = "^0.5.0"
Flask-Bcrypt = "^0.7.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pylint = "^2.6.0"
black = "^20.8b1"
pylint-flask-sqlalchemy = "^0.2.0"
pylint-flask = "^0.6"
pytest-xdist = "^2.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
